Merge from master

+++ geodirectory-functions/ajax_handler_functions.php
@@ -329,8 +329,13 @@ function geodir_ajax_handler()
 
                         global $current_user;
                         get_currentuserinfo();
-                        $post_type = get_post_type($_REQUEST['pid']);
-                        $lastid = wp_delete_post($_REQUEST['pid']);
+
+                        if(get_option('geodir_disable_perm_delete')){
+                            $lastid = wp_trash_post($_REQUEST['pid']);
+                        }else{
+                            $lastid = wp_delete_post($_REQUEST['pid']);
+                        }
+
                         if ($lastid && !is_wp_error($lastid))
                             wp_redirect($_SERVER['HTTP_REFERER']);
